Fixed professor grade input not backspacing the 0, default grade is null instead of 85, fixed date format in student dashboard

+++ Spring-App/src/main/java/com/example/demo/models/Submission.java
@@ -17,7 +17,7 @@ public class Submission {
     private String fileType;
     private String status;
     private String feedback;
-    private double grade;
+    private Integer grade;
     private String documentId;
     private Date submissionDate;
     private Date lastModified;
@@ -123,11 +123,11 @@ public class Submission {
         this.feedback = feedback;
     }
 
-    public double getGrade() {
+    public Integer getGrade() {
         return grade;
     }
 
-    public void setGrade(double grade) {
+    public void setGrade(Integer grade) {
         this.grade = grade;
     }
